Check for thread delete permission (controller level)

+++ app/Http/Controllers/ThreadsController.php
@@ -68,6 +68,14 @@ class ThreadsController extends Controller
 
     public function destroy($channel, Thread $thread)
     {
+        if ($thread->user_id != auth()->id()) {
+            if (request()->wantsJson()) {
+                return response(['status' => 'Permission denied'], Response::HTTP_FORBIDDEN);
+            }
+
+            abort(Response::HTTP_FORBIDDEN, 'You do not have permission to do this');
+        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
